requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
yfinance==0.2.43
pandas==2.2.2
numpy==2.0.2
openpyxl==3.1.5
XlsxWriter==3.2.0
requests==2.32.3
